fix compiler blowup error
