+++ src/main/scala/ammonite/kernel/LogMessage.scala
@@ -1,7 +1,21 @@
 package ammonite.kernel
 
+import java.io.{StringWriter, PrintWriter}
+
 sealed trait LogMessage {
   val msg: String
+}
+
+object LogMessage {
+  def fromThrowable(t: Throwable): LogError = {
+    val sw = new StringWriter();
+    val pw = new PrintWriter(sw);
+    t.printStackTrace(pw);
+    val msg = sw.toString();
+    sw.close()
+    pw.close()
+    LogError(msg)
+  }
 }
 
 final case class LogError(msg: String) extends LogMessage
